FIX: When KEYMAT not configured, None is passed to CFFI causing errors
When the application is deployed to an ElasticBeanstalk environment and the SR_ENVIRONMENT variable is set,
the application looks for the presence of two env vars EQ_RRM_PUBLIC_KEY and EQ_SR_PRIVATE_KEY. If these
are set to None then the application fails, as it can't initialise a Decoder correctly for JWT usage.

This commit adds a check for the presence of those two settings, raising an OSError (closest error type ish) to
highlight the error as soon as it appears.

This error (and all errors) are currently surpressed by incorrect HTTP 500 error handlers that hampered finding this bug.
The next commit will address this issue.