Added McCabe index to CI tests
To help us write and refactor our codebase to reduce complexity, we're implementing a CI test for the McCabe index (https://en.wikipedia.org/wiki/Cyclomatic_complexity). The aim of this check is to minimise the complexity of any individual function, which has been shown to be correlated to numbers of bugs in a codebase.

We're initially taking the baseline of 10 as our limit.