Add dynamic theme support
This commit adds the ability for a survey schema to define a theme and for that theme (if it exists) to be used to
display that survey in the runner.

We create a new theme 'star_wars' which is a stub for a social style theme.

Added a theme selection to the Star wars kitchen sink questionnaire.