Alter connection logic to rabbitmq to handle failover
As part of the removal of the ELB in our architecture, we've changed how we connect and handle rabbitmq clusters. This commit
changes how we connect, adding the failover to a secondary backup RabbitMQ server in the event of not being able to connect to the
prime.

In the event that the secondary fails, we generate a 5XX server error and log to the error logger.