Driving functional tests using page models
The functional tests were going down a path where it would have started
to get harder to manage if changes to survey schema occured because we
would potentially have to make changes in multiple places. We have
introduced page models which should be an abstraction on top of a
webpage e.g. the landing page.

Also renamed the wdio test folder to reflect the type of tests rather
than the technology implementing them