eq-364 Other option for radio/checkboxes (#475)
Added feature to JSON survey to allow other option field to be used on checkboxes and radio buttons. This allows the user to enter in their own value if this option is selected. The value of the other field is posted using the same name as the multi choice answer and provided as an array. The backend extracts the user entered value instead of the other label.