Fixing question formatter unit tests
The unit tests were failing due to changes to the summary model changing
which now using the questionnaire state utility method to find the
question and answer schema. The questionnaire state was `None` causing
invocation exceptions.

The test has been refactored to use mocking to write less code to
achieve the fix