Move static content up a directory
Generated static content was being created in an 'app/static' folder. This caused problems in the Docker build when people had already compiled the assets locally (the app folder is mounted as a volume in the container). Static content is now created in a 'static' folder that does not get mounted as a volume. It also creates a clean separation between version controlled content and generated content.