Fixing SQLAlchemy transaction handling
We saw a error in production which suggested a transaction had already
been closed `sqlalchemy.exc.InvalidRequestError: This Session's
transaction has been rolled back due to a previous exception during
flush`. On further investigation this was because of a duplicate
questionnaire PK when committing a transaction. Further down the line
the application tried to make use of the session which due to the error
meant we had no transaction to use as it was left in a dirty state
without being rolled back. This caused a stacktrace to be displayed on
simulation.

* Removing db_session when application dies and at the end of a request
  as per [flask documentation]
  (http://flask.pocoo.org/docs/0.11/patterns/sqlalchemy/#declarative).
* Catching errors on commit and rolling back then re-raising error.
* Removing unused interface
* Logging tx_id against errors

Further reading: On investigating different transaction handling
strategies one suggestion which seems quite applicable in our scenario
was to map the transaction to the http request which makes sense when
following the `unit of work` pattern. The benefits of this would be we
wouldn't have to manually handle our transactions for all of the
database code and we'd have one place to handle it thus one area to test
regarding transactions.