Add conditional routing to a previous group location ========================================
Adds conditional routing to a previous location. Allows us to support the routing of users to the household composition screen in census, based on the answer to a question. This is handled by adding a route to the routing path and removing the users current answer to this question (which would otherwise prevent the location path being calculated by navigator).

Additionally, I've added a pluralize jinja filter to support person/people is/are based pluralization of text, which will need some work in another story to evaluate a more appropriate approach.