Add Household composition real time (client side) changes
Enables the "realtime" adding/removing of household members in the household composition question.

In order to guarantee there will be no side effects for future dependent groups (relationships, household member details) we
remove all answers for all household members on POST of the household composition screen. This is not the greatest user experience,
but future stories will improve on this and only wipe what is necessary.

This has modified the functionality of the remove method in the answer
store to remove any answers identified by filter criteria. The original
remove method has been renamed to remove_answer.