Modify sexual id routing to use both metadata and answer response ================================================================= This change updates the routing around sexual identity question to be informed by both the metadata flag coming from respondent home and the answer to the over 16 question.
As part of this it has been necessary to update all when rules to
supply conditions as lists and all behaviour that uses them to
evaluate every one of the conditions supplied.

Additionally, we were incorrectly converting the sexual identity
flag supplied by respondent home to a string value, so I have had to
modify the routing rules/evaluation to work against a boolean value.

I've also got rid of the SchemaCondition/When classes and just used
the rule evaluation from dicts throughout.

It's also been necessary to simplify _build_path slightly as our
tests started failing due to the computational complexity of it
being too high!