Fix repeating group bug for census
Repeating answers will repeat if the answer which determines the repeat
was originally answered but if the user then changes their mind to go
down a path where the answer which determines the repeat will no longer
appear on. The expected result is for those repeating answers to not be
repeated. The actual result is that the repeating answers are repeated.

The scenario this occurs under is:
Given I have added a person in my household, When I change my mind and
they should be visitors, Then I do not have to answer household member
details

The short term fix for this is to remove the answer which determines the
repeat when the user goes down the other path which is consistent with
what we do when a user changes their answer which determines the number
of repeats.

The ideal fix going forward would be to check if the answer which
determines the repeating answers is on the path that the user is being
followed.

[trello](https://trello.com/c/vnjf9dR9/813-eq-survey-runner-727-when-clicking-previous-after-adding-household-members-the-reporting-groups-still-repeat-issue)
[trello for the real fix](https://trello.com/c/QXkI2J6i/298-repeating-groups-should-only-be-evaluated-if-the-answer-which-determines-the-repeating-groups-in-on-the-path)