Add navigation capability
Adds navigation capability into survey runner, tested and working for Census use case. Each group appears in the navigation unless it has a repeating rule which hasn't been completed then it is hidden. If a group is repeating and the feeding question has been answered it then repeats the required amount. Navigation when clicking 'Save and Continue' is always linear, unless you reach the end, and it then goes to the last completed block. The previous link is not present on the first block of each group.

Work was completed by @LJBabbage but the original branch was in rebase hell.