Adding skip_conditions for groups
We need to be able to skip the visitors completed interstitial page in
the census household questionnaire. It lives in a separate group to the
visitor questions therefore it won't be skipped by the repeat rule.

This change implements skip_condition on a group to solve that.

Fixes #740 and Fixes
https://trello.com/c/wGkp3T52/291-remove-need-to-update-state-in-navigator

[Open bug](https://trello.com/c/17GX3qtI/331-a-skip-condition-on-the-last-group-will-causes-an-error)