Support `or` rules for group skip_condition
We need to be able to support `or` rules for the census visitors
interstitial page so that it is skipped when the number of visitors is 0
or skipped via navigation. We have changed the skip_condition
configuration to take an array of when elements. Adding more than one
when element creates a `or` rule like so.

```
"skip_condition": [
                {
                    "when": [
                        {
                            "id": "overnight-visitors-answer",
                            "condition": "equals",
                            "value": "0"
                        }
                    ]
                },
                {
                   "when": [
                       {
                           "id" : "overnight-visitors-answer",
                           "condition": "not set"
                       }
                   ]
                }
            ],
```