Re-add Apache WSGI config fix
The error `extern "Python": function Cryptography_pem_password_cb() called, but @ffi.def_extern() was not called in the current subinterpreter.  Returning 0.` started occurring on one of the Elastic Beanstalk ec2 instances after the Apache fix was removed in #870. We need to be able to reliably reproduce the error before we attempt to remove the Apache fix again.