Make session timeout configurable via the schema
The session timeout was previously made configurable by using session
storage, but this won't work when using phantomjs for testing. This
change also allows the timeout to be configured per survey if required.

As part of this pull request the grace period has been changed so that
it is added to the server side session rather than subtracted from the
timeout period sent to the client. This is so that when the timeout is
defined in the schema, that is the timeout the client experiences
before being timed out.