Track jti claim to prevent repeat attacks
JWT rfc defines a token which can be used to prevent repeat attacks.
e.g. login using the same token twice by using the same JWT token before
the JWT token has expiried.
The token is defined here: [jti](https://tools.ietf.org/html/rfc7519#section-4.1.7)

If the jti token is present in the JWT payload we store the jti token in
a postgres table called used_token.
If the same token is seen more than once the authentication request is
rejected as invalid.
If the jti token is not present we do not protect against repeat
attacks. This puts the responsibility on the upstream system.

The tokens are stored in a postgres table and old tokens can be cleared
out by referencing the timestamp column. i.e. `where timestamp > '1
month old'`