Escaping quotes and backslash in answers
We use jinja to inject strings into the questionnaire schemas however if
quotes or backslashes are injected it will break.

Quotes nor backslashes can be in JSON without being escaped. Both could
get into the JSON via user answers if they are not escaped before hand.

This fix is a minimal fix for the problem described.

* Escape business names to make 'json safe'