Refactored the way user input is handled
All input encoding has been pushed down to the schema generation level
where the answer store is accessed directly.

As a result Jinja autoescaping has been turned off at the theme
template level where the generated schemas are consumed. There was a
bunch of test refactoring required as a result but we should be in a
better place when we come to revisit the schema generation in the future.