1151 Schema Generation from block files + manifests
There's a new script which will generate schemas from mainfests found under data/sources.
The script has been added to the build process and there are manifests + blocks for all the
refreshed RSI and MCI form types.

Edits to the refreshed RSI and MCI schemas have also been made as during review of the
generated schemas we spotted that some of the new guidance content had the wrong structure.