Users can preview the info needed for a survey
UI Changes

- Accordion reworked for previewing survey info
- MWSS theme created to cater for new start button location

Content Changes

MWSS now has an overview of the information users will need to
answer the survey.

Schema Chnages

Blocks can now have a body property and are no longer allowed to have
'basis-for-completion' or 'information-to-provide' properties.

Blocks can still have 'title' and 'description' properties as I'm not
sure whether changing *all* blocks to have a body property where
titles and descriptions live is something we want to do.

The id of 'Basic' block body objects are now used to populate the div
id in the introduction template.