Fixed CSS minification + source maps
After some poking around (and upgrading the gulp sass plugin) it looks
like a problem with the globs + postCSS + cssnano was resulting in a <no
source> file being added to the gulp styles task. Seems like the
likely root cause so I've fixed it.

Source maps are now generated in all environments.