No JS Feedback form
Leverages flask_wtf csrf protection. This resulted in a bit of a
refactor of the questionnaire view as I tried to reuse the same form
validation check.

There is now a form helper which is shared between the feedback and
questionnaire views. The GET and POST actions for blocks in the
questionnaire blueprint have also been heavily re-worked as I
integrated the use of the form helper.

Human readable logging is now the default when runner is using the dev
settings.

Added decorators for template rendering + path helper

Added a Template helper, which is just a collection of decorators and
is being used in the questionnaire and feedback views. The granularity
of each decorator hasn't been fully thought through apart from rough
adherence to the Single Responsibility Principle. Worth evaluating as
we work with them in the future.

path_finder now uses Werkzeugs LocalProxy pattern and has been moved
to its own helper file.

Amended integration test base class to proxy kwargs. I needed to be
able to set a Referer header to test the feedback message generation
properly.

CSRF protection is now the default for the app. If you don't want it
you need to code an exemption. You can exempt at the blueprint and
view level. Currently the Flush and Session blueprints are the only
ones exempted.

Session related views/actions have been migrated from the
Questionnaire blueprint to the Session blueprint. This has changed the
endpoints for expiring sessions, handling sign outs, etc. This is a
more logical grouping than having those endpoints tagged onto the end
of a questionnaire URL.

Amended 'multiple surveys' functional test. It's now a stale survey
test. This is due to the now global CSRF protection kicking in as in
this scenario the CSRF token in the survey that is opened first
becomes stale.

Removed the javascript constants for the session urls which are now
hard coded in the timeout js module. This was as a result of the
session endpoints being moved into the Session blueprint (there's no
need for a dynamic prefix anymore).

Replaced an RSI specific functional test with a generic, and more
complete, test suite that describes the expected behaviour when a user
launches two surveys at the same time.

New functional tests now use headless chrome