Only re-generate routing_path on changes to the answer_store
The routing path was being generated about 7 times per request.
This PR means that on a GET the routing_path is generated once and on a POST it is generated twice. The answer store object knows about (can generate a hash, with new method get_hash, of) the latest answers.

The questionanaire views contained many scattered calls to the path finder methods to the extent that it was not obvious where a routing path was required. This adds a decorator for views which passes in the full path as the first parameter to the decorated view which can then be used for each of the helper functions making it more clear when a path is calculated in a request.