Add collapsible inline feedback form
Now that the feedback form template appears in every questionnaire
page the attribute used for selecting the submit button needs to be
different to the submit button for a question.

The POST payload for the feedback form is checked for a 'redirect'
value. If the value is the string 'false' the request doesn't get
redirected to the thankyou page.

Added a functional test for the inline feedback. This includes the
addition of a FeedbackForm class which is used by the new test and the
existing static feedback functional test.

GA events and tests added for "opening" and "closing" the feedback form and submitting it.

As requested, the feedback blueprint now logs when it processes requests in a
similar fashion to the questionnaire blueprint.