Fix 500 error on unauthenticated thank you page request
After the recent changes in #1222 a HTTP 500 error was being generated
when visiting a thank-you page URL without a cookie. This was due to
path finder trying to access the questionnaire store without any user
context. The fix applies the @login_required decorator to the path
finder global. This generates a 401 error when called without a user
session.

Fixes: #1277