Add script for generating requirements.txt file
'pipenv lock -r' outputs a list of dependencies suitable for piping to requirements.txt file but it tends to include all (including dev) dependencies. This is expected behaviour and unlikely to change in Pipenv project but we can work around this by using the Pipenv API to generate our own requirements.txt file (with hashes).

Fixes: #1324