Fix test_numbers.json schema to pass validation.
Adds a min/max to the question that sets the min/max values for the subsequent range questions.
This essentially prevents the situation where a user may input a max value that is smaller than the min value. Ideally we would validate that max is greater than min on the set-min-max-block question but this isnt necessary to satisfy any requirements at the moment.