Don't save state in new format as it won't work during deployment
As there is a chance of requests going to a new version of the app
and then back to an old version during deployment, we can't persist
any format changes to state in the next release. This release should
however have all the code required to understand the new state
formats, which are:
  - Questionnaire store compressed via snappy, and base64
    encoding and json data wrapper removed
  - Removal of base64 encoding for session data
  - View submission data in answer map format

So, to be clear, anything written (questionnaire state, session or
view submission data) in this version of the app should be readable
by the currently released version (v2.73.0).