Rename QPSES schema files.
QPSES is being treated as 3 distinct surveys from a RAS/RM perspective and two of those need to use the same form type (0002).

In order for EQ to be able to support two QPSES questionnaires with the same form type the survey ID needs to be included in the EQ_ID to allow for unique file names.
Since the schema files are a combination of EQ_ID and form type.