Update PyYAML to later version.
A high serverity security vulnerability is being reported in one of the project dependencies.
See CVE-2017-18342 for further details. https://nvd.nist.gov/vuln/detail/CVE-2017-18342

This change updates the version of PyYAML to a patched version to mitigate the issue.
The dependency can be changed back to "*" in the Pipfile once a released version has been made available on PyPI.