Fix error when feedback form submitted after questionnaire submitted and session ended.
This change aims to fix #1809 where it is possible to access the feedback page, even when the user no longer has a valid session.
The change ensures that a 401 response is returned whenever it is not possible to build the metadata context because of an invalid user session.
Preventing feedback form from being submitted if session expired. Also reverted previous changes to .get() metadata values