Escape answers when rendered on a page (#2331)
Escapes answers when used in placeholders, answer fields 
and summaries.

Escaping of answers was mistakenly removed in 
https://github.com/ONSdigital/eq-survey-runner/pull/2040.

For some background, autoescaping was turned off in
https://github.com/ONSdigital/eq-survey-runner/commit/159ecadfb97772cb813faf47f75a1e28852a7161 

Before autoescaping was turned off `safe` was used throughout 
the templates to turn off escaping for strings that could 
contain HTML. Most strings in templates used this approach. 
This lead to safe being used everywhere without any thought 
about whether the contents had been escaped or not. By 
making the choice explicit it made it clearer what should 
be escaped.