Handle languages for Census schemas (#2365)
The original intention was to add the allowed languages to the schema. This hasn't been done as the language currently needs to be known before loading the schema. Instead there is a hardcoded mapping in code of schema name to allowed languages. We will need to look at how we want to solve this going forward.

`en_US` has been used for Ulstér Scotch as there is no language code in the CLDR (http://cldr.unicode.org) for it and this is required for survey runner to work. This introduces two issues:
- month names will be in English
- placeholders that have possessive names will have an `'s` added