Switch from persistent to session cookies
After investigation into production issues it has been identified
that some browsers (predominantly Edge) seem to drop our session
cookie randomly. Although we still don't know why this happens
(debugging has shown that a valid cookie is sent on the request
previous to the one that errors), changing from a permanent cookie
(one with an expiry time set) to a session cookie (one that expires
when the browser is closed) seems to fix the issue.

Arguably the cookie should always have been a session one. The
server side session times out after the timeout period (default is
45 minutes) and keeping the cookie client side means we will still
have context like the theme or an account service url that are
useful when rendering error screens. The cookie was set to expire
prior to us having server side session expiry.