🛠 Allow GA cookies to have single or double quotes
Refactors out analytics allowed checking code in runner into its own
function so can be modified in one place if need be. This function
parses the cookie as JSON, first replacing single with double quotes in
order to allow this to happen.

It now works with both double and single quotes.